+++ src/unicodedb/types.nim
@@ -8,19 +8,19 @@ import types_data
 
 export UnicodeTypeMask
 
-proc unicodeTypes*(cp: int): int =
-  ## Return types for a given code point.
-  ## Use `contains` to retrieve a single type
-  assert cp <= 0x10FFFF
-  let
-    blockOffset = int(typesOffsets[cp div blockSize]) * blockSize
-    idx = typesIndices[blockOffset + cp mod blockSize]
-  result = typesData[idx]
-
 proc unicodeTypes*(cp: Rune): int {.inline.} =
   ## Return types for a given code point.
   ## Use `contains` to retrieve a single type
-  unicodeTypes(cp.int32)
+  assert cp.int <= 0x10FFFF
+  let
+    blockOffset = (typesOffsets[cp.int div blockSize]).int * blockSize
+    idx = typesIndices[blockOffset + cp.int mod blockSize]
+  result = typesData[idx]
+
+proc unicodeTypes*(cp: int): int {.deprecated.} =
+  ## **Deprecated since version 0.3.0**;
+  ## Use ``unicodeTypes(Rune)`` instead.
+  unicodeTypes(cp.Rune)
 
 proc contains*(ut: int, utm: UnicodeTypeMask): bool =
   ## Check if the given type mask is